requirements.txt
asgiref==3.8.1
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
numpy==2.2.6
packaging==25.0
pandas==2.2.3
psycopg2==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tzdata==2024.2
whitenoise==6.9.0
